Improved keybind integration with LCCore + whitespace

+++ src/main/java/me/fril/regeneration/common/EntityFobWatch.java
@@ -214,8 +214,8 @@ public class EntityFobWatch extends Entity {
 		if (flag || this.ticksExisted % 25 == 0) {
 			if (this.world.getBlockState(new BlockPos(this)).getMaterial() == Material.LAVA) {
 				this.motionY = 0.20000000298023224D;
-				this.motionX = (double) ((this.rand.nextFloat() - this.rand.nextFloat()) * 0.2F);
-				this.motionZ = (double) ((this.rand.nextFloat() - this.rand.nextFloat()) * 0.2F);
+				this.motionX = (this.rand.nextFloat() - this.rand.nextFloat()) * 0.2F;
+				this.motionZ = (this.rand.nextFloat() - this.rand.nextFloat()) * 0.2F;
 				this.playSound(SoundEvents.ENTITY_GENERIC_BURN, 0.4F, 2.0F + this.rand.nextFloat() * 0.4F);
 			}
 		}
@@ -228,9 +228,9 @@ public class EntityFobWatch extends Entity {
 			f = underState.getBlock().getSlipperiness(underState, this.world, underPos, this) * 0.98F;
 		}
 		
-		this.motionX *= (double) f;
+		this.motionX *= f;
 		this.motionY *= 0.9800000190734863D;
-		this.motionZ *= (double) f;
+		this.motionZ *= f;
 		
 		if (this.onGround) {
 			this.motionY *= -0.5D;
